Add the ability to explicitly supply a default CNI plugin to run when no annotation is found.
This PR is based off https://github.com/Huawei-PaaS/CNI-Genie/pull/30.
In addition to adding a default in the genie netconf file. It fixes an
issue with this base PR where it would Unmarshall another CNI plugin
into a Genie specific NetConf struct then Marshal it back out to JSON.
This caused the delegate plugin to fail usually because it was being
passed JSON fields that were empty. The solution was to pass the
delegate CNI plugin's netconf config into the ipam call, without the
marshall/unmarshall steps.